Added scrolling, in which map data is read from progmem
For testing, progmem is imitating the SD card.  The next phase
will be to move the data to SD and change the progmem reads to
SD card reads.

If the map buffer is made smaller, I draw lines that roughly
indicate the viewport that would work with the smaller buffer.
This shows how tile data is loaded outside the viewport.  It's an
interesting effect.