Switched code to load from SD card, using petit_fatfs
This SD library is in a a way a worst-case scenario for performance
because it doesn't allocate a 512-byte buffer like other SD
libraries do.  This means that reading column data is very
expensive.  When testing this, scrolling left and right causes lag
while scrolling vertically is smooth.  This makes me wonder if a
different SD library might result in better performance...